move estimateFeeRate and getFeesRateBlockcypher methods to common/utils/coin
